ai_copilot-main/backend/requirements.txt
vapi_python
groq
fastapi
uvicorn
python-dotenv
pydantic
httpx
websockets
python-multipart
modal 